[UCR] Fix issue with creating bundles

+++ create_bundle.py
@@ -38,7 +38,7 @@ if __name__ == "__main__":
                     archive.open(archive_file), encoding="utf-8"
                 ) as io_wrapper:
                     arff, _metadata = loadarff(io_wrapper)
-                    arr = np.array(arff.tolist())
+                    arr = np.array(arff.tolist()).astype(np.float32)
                     if NO_MISSING and not np.any(np.isnan(arr)):
                         continue
                     np.save(
